rename: Carousel 컴포넌트 StoreCarousel로 이름 변경
