+++ src/pages/store/index.js
@@ -1,5 +1,4 @@
-import Carousel from 'components/Carousel';
-import Image from 'next/image';
+import StoreCarousel from 'components/StoreCarousel';
 
 import React from 'react';
 
@@ -7,7 +6,7 @@ export default function Store() {
   return (
     <>
       <h1 className="title">Store</h1>
-      <Carousel />
+      <StoreCarousel />
     </>
   );
 }
